AI-TICKET-TRIAGE/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
pydantic==2.9.0
pydantic-settings==2.5.0
sqlalchemy==2.0.35
aiosqlite==0.20.0
openai==1.50.0
tiktoken==0.7.0
faiss-cpu==1.8.0
langchain==0.3.0
langchain-openai==0.2.0
langchain-community==0.3.0
streamlit==1.38.0
plotly==5.24.0
altair==5.4.0
scikit-learn==1.5.0
pandas==2.2.0
numpy==1.26.0
python-dotenv==1.0.1
httpx==0.27.0
loguru==0.7.2
pytest==8.3.0
pytest-asyncio==0.24.0